refactor: MessageBody 구현체 messagebody로 패키징
