+++ src/main/java/http/request/messagebody/HttpFormDataMessageBody.java
@@ -1,5 +1,6 @@
-package http.request;
+package http.request.messagebody;
 
+import http.request.MessageBody;
 import http.request.bodyparser.FormDataParser;
 
 import java.util.Map;
