+++ src/test/java/http/request/ContentTypeTest.java
@@ -1,5 +1,6 @@
 package http.request;
 
+import http.request.messagebody.HttpFormDataMessageBody;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
